fixes frozen scrolling after drawer close
